Add files via upload
Hey everyone! This project is telling you to tell about prevention of Covid. You haven't wear mask and all the people around you are corona positive. You need to reack to your mother without having the virus. Try to reach and be safe. 
(note:- The instructions are given in game.)